Adding a relation between Category and Question
