+++ src/main/kotlin/com/theworkshopvlc/chupapuntes/categories/CategoriesController.kt
@@ -1,6 +1,6 @@
 package com.theworkshopvlc.chupapuntes.categories
 
-import com.theworkshopvlc.chupapuntes.categories.domain.Category
+import com.theworkshopvlc.chupapuntes.categories.domain.entities.Category
 import com.theworkshopvlc.chupapuntes.categories.domain.usecases.GetAllCategories
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
