+++ src/main/kotlin/com/theworkshopvlc/chupapuntes/categories/persistence/CategoriesDAO.kt
@@ -1,6 +1,6 @@
 package com.theworkshopvlc.chupapuntes.categories.persistence
 
-import com.theworkshopvlc.chupapuntes.categories.domain.Category
+import com.theworkshopvlc.chupapuntes.categories.domain.entities.Category
 import org.springframework.data.repository.CrudRepository
 
 interface CategoriesDAO : CrudRepository<Category, Long>
