+++ src/main/kotlin/com/theworkshopvlc/chupapuntes/questions/model/entities/Question.kt
@@ -1,9 +1,11 @@
 package com.theworkshopvlc.chupapuntes.questions.model.entities
 
+import com.theworkshopvlc.chupapuntes.categories.domain.entities.Category
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType
 import javax.persistence.Id
+import javax.persistence.ManyToMany
 
 @Entity
 data class Question(
@@ -13,6 +15,9 @@ data class Question(
   val title: String,
 
   val description: String,
+
+  @ManyToMany(mappedBy = "categories")
+  val categories: Set<Category> = emptySet(),
 
   val createdAt: Long = System.currentTimeMillis()
 )
